chore: rename typescript-specific files
