+++ src/fuzzer/analysis/typescript/TypescriptProgram.ts
